Use intptr_t for stack registers

+++ mkt_stdlib.c
@@ -15,8 +15,8 @@ static u64 gc_allocated_bytes = 0;
 static const unsigned char RV_TAG_MARKED = 0x01;
 static const unsigned char RV_TAG_STRING = 0x02;
 static const unsigned char RV_TAG_INSTANCE = 0x04;
-static i64* mkt_rsp = NULL;
-i64* mkt_stack_top = NULL;
+static intptr_t* mkt_rsp = NULL;
+intptr_t* mkt_stack_top = NULL;
 
 #define READ_RSP() __asm__ volatile("movq %%rsp, %0" : "=r"(mkt_rsp))
